Create README.md
owards the end of 1997 Geoff Banks, then Mayor of Cowes on the Isle of Wight, called a public meeting at Northwood House to draw attention to the need to perpetuate the memory of the events of May 1942, when ORP Błyskawica and her ship's company had defended Cowes from enemy air attack. Those present at the meeting agreed to form an organisation with this objective. The Friends of the Błyskawica Society was constituted in March 1998 and its first Annual General Meeting was held in January 1999 under the Chairmanship of John Power. The other Committee members at that time were Ken Sinnicks, Otton Hulacki, Jaycey Hulacka, Irena Wells, John Coueslant, Frank Aze, Wladek Przygonski and Malcolm Turner. Geoff Banks was made President.In 2008 the organisation became the “Friends of the ORP Błyskawica Society”. (ORP= Okret Rzeczypospolitej Polskiej= Ship of the Republic of Poland).

The Society’s aims, as stated in its Constitution, are to “raise awareness of the connection with Cowes and the role played in the 2nd World War by the Polish Naval Ship “BLYSKAWICA” and to raise funds to help with the maintenance and preservation of the Ship